changed

+++ src/discount_cupon/entities/discount_cupon.entity.ts
@@ -3,7 +3,7 @@ import { Column, DataType, Model, Table } from "sequelize-typescript";
 @Table({tableName: 'discount_cupons', timestamps: false})
 export class DiscountCupon extends Model<DiscountCupon> {
     @Column({
-        type: DataType.SMALLINT,
+        type: DataType.BIGINT,
         unique: true,
         autoIncrement: true,
         primaryKey: true,
